linker N: Improve gdb support
With default bionic's implementation, the list of loaded libraries
for GDB is reset when the first Android lib is loaded.
With libhybris, we want to get both Android and Glibc loaded libraries.

However, we shouldn't simply link our "link_map" pointers in the
current list, because glibc's "link_map" object shouldn't point to
our incompatible "link_map" objects. To avoid that:
 1. make a copy of the initial list, to avoid modifying glibc's objects
 2. prepend our objects to the list, instead of appening it

 That way glibc's objects are never modified, and glibc's linker won't
 try to manipulate objects that haven't been initialized by him.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>